+++ packages/data-explorer/src/lib/client.ts
@@ -1,0 +1,14 @@
+import axios from 'axios'
+
+const client = axios.create({
+  baseURL: '/',
+  timeout: 5000,
+  validateStatus: function (status) {
+    if (status === 401) {
+      window.location.href = '/login'
+    }
+    return status >= 200 && status < 300
+  }
+})
+
+export default client